% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alias}
\alias{alias}
\alias{es_alias_create}
\alias{es_alias_delete}
\alias{es_alias_exists}
\alias{es_alias_get}
\alias{es_aliases_get}
\title{Elasticsearch indices APIs}
\usage{
es_alias_get(index = NULL, alias = NULL, ignore_unavailable = FALSE,
  callopts = list())

es_aliases_get(index = NULL, alias = NULL, ignore_unavailable = FALSE,
  callopts = list())

es_alias_exists(index = NULL, alias = NULL, callopts = list())

es_alias_create(index = NULL, alias, routing = NULL, filter = NULL,
  callopts = list())

es_alias_delete(index = NULL, alias, callopts = list())
}
\arguments{
\item{index}{A comma-separated list of index names}

\item{callopts}{Curl args passed on to httr::POST.}

\item{features}{A comma-separated list of features. One or more of settings, mappings,
warmers or aliases}

\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then raw JSON.}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{...}{Further args passed on to elastic search HTTP API as parameters.}

\item{type}{Document type}

\item{id}{Document id}

\item{fields}{Fields to add.}
}
\description{
Elasticsearch indices APIs
}
\examples{
\donttest{
# Retrieve a specified alias
es_alias_get(index="plos")
es_alias_get(alias="*")
es_aliases_get()

# Check for alias existence
es_alias_exists(index = "plos")

# Create/update an alias
es_alias_create(index = "plos", alias = "tables")

# Delete an alias
es_alias_delete(index = "plos", alias = "tables")
}
}
\author{
Scott Chamberlain <myrmecocystus@gmail.com>
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/indices-aliases.html}
}

