% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsl.R
\name{bool}
\alias{bool}
\title{Define bool query for the DSL}
\usage{
bool(.obj = list(), ...)
}
\description{
Define bool query for the DSL
}
\examples{
\dontrun{
index("shakespeare") \%>\%
   bool(must = list(term=list(user="kimchy")))
bool(must_not = list(term=list(user="kimchy")))
bool(should = list(list(term=list(tag="sometag")), list(term=list(tag="someothertag"))))

bool(must = list(term=list(user="kimchy")),
     must_not = list(term=list(user="kimchy")))

}
}

