% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Cat.R
\docType{data}
\name{Cat-class}
\alias{Cat-class}
\alias{Cat}
\title{Cat class}
\arguments{
\item{verbose}{(logical) If \code{TRUE} (default) the url call used printed
to console}

\item{index}{(character) Index name}

\item{fields}{(character) Fields to return, only used with \code{fielddata}}

\item{h}{(character) Fields to return}

\item{help}{(logical) Output available columns, and their meanings}

\item{bytes}{(logical) Give numbers back machine friendly. Default: \code{FALSE}}

\item{parse}{(logical) Parse to a data.frame or not. Default: \code{FALSE}}

\item{...}{Curl args passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
holds all cat API methods
}
\details{
xxxx
}
\examples{
\dontrun{
(x <- connect())
x$ping()
z <- Cat$new(x)
z
z$conn
z$help()
z$aliases()
z$allocation()
z$count()
z$segments()
z$health()
z$indices()
z$master()
z$nodes()
z$nodeattrs()
z$pending_tasks()
z$plugins()
z$recovery()
z$thread_pool()
z$shards()
z$fielddata()
}
}
\keyword{datasets}
