% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cdbriver.r
\name{cdbriver}
\alias{cdbriver}
\alias{cdbriver_auth}
\alias{cdbriver_index}
\title{Interact with CouchDB via Elasticsearch river plugin.}
\usage{
cdbriver_auth(user = NULL, pwd = NULL, dbname,
  endpoint = "http://localhost", port = 9200)

cdbriver_index(dbname, endpoint = "http://localhost", port = 9200,
  what = "start")
}
\arguments{
\item{user}{Username}

\item{pwd}{Password}

\item{dbname}{Database name. (character)}

\item{endpoint}{the endpoint, defaults to localhost (http://127.0.0.1)}

\item{port}{port to connect to, defaults to 9200}

\item{what}{One of start (default) of stop.}
}
\description{
Interact with CouchDB via Elasticsearch river plugin.
}
\examples{
\donttest{
# Authenticate
cdbriver_auth(user="foo", pwd="bar")

# Start or stop indexing a CouchDB database using Elasticsearch.
devtools::install_github("sckott/sofa")
library('sofa')
sofa_createdb(dbname='mydb')
cdbriver_index(dbname='mydb')
cdbriver_index(dbname='mydb', what='stop')
}
}
\references{
See docs for the Elasticsearch River plugin
\url{https://github.com/elasticsearch/elasticsearch-river-couchdb/blob/master/README.md}
that lets you easily index CouchDB databases.
}

