% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/elastic-package.r
\name{elastic-defunct}
\alias{elastic-defunct}
\title{Defunct functions in elastic}
\description{
\itemize{
 \item \code{\link{mlt}}: The MLT API has been removed, use More Like This Query
 via \code{\link{Search}}
}
}

