% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{elastic}
\alias{elastic}
\alias{elastic-package}
\title{elastic: An Elasticsearch R client.}
\description{
elastic: An Elasticsearch R client.
}
\section{About}{


This package gives you access to local or remote Elasticsearch databases.
}

\section{Quick start}{

}

