% Generated by roxygen2 (4.0.1): do not edit by hand
\name{elastic_POST}
\alias{elastic_POST}
\title{Wrapper around httr::POST for this package}
\usage{
elastic_POST(path, index = NULL, type = NULL, metric = NULL, node = NULL,
  clazz = NULL, data=NULL, raw, ...)
}
\arguments{
\item{path}{Elasticsearch API endpoint path}

\item{index}{Elasticsearch index}

\item{type}{Elasticsearch type}

\item{clazz}{Class to outupt}

\item{raw}{Raw JSON results as string}

\item{data}{stringified post object, -d from curl examples}

\item{...}{Further args passed to Elasticsearch}
}
\description{
Wrapper around httr::POST for this package
}

