% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsl.R
\name{elasticdsl}
\alias{bool}
\alias{bool_}
\alias{boosting}
\alias{boosting_}
\alias{common}
\alias{common_}
\alias{elasticdsl}
\alias{filter}
\alias{geoshape}
\alias{geoshape_}
\alias{index}
\alias{n}
\alias{print.index}
\alias{range}
\alias{range_}
\title{The elastic DSL}
\usage{
index(index, type = NULL, what = "mappings", ...)

\method{print}{index}(x, ...)

range(.obj = list(), ..., boost = 1, time_zone = NULL, execution = NULL,
  cache = FALSE)

range_(.obj = list(), ..., .dots)

bool(.obj = list(), ...)

bool_(.obj = list(), ..., .dots)

geoshape(.obj = list(), ..., field = NULL)

geoshape_(.obj = list(), ..., .dots, field = NULL)

n(x)

filter()

boosting(.obj = list(), ..., negative_boost = NULL)

boosting_(.obj = list(), ..., .dots, negative_boost = NULL)

common(.obj = list(), field, query = NULL, cutoff_frequency = NULL,
  low_freq_operator = NULL, minimum_should_match = NULL)

common_(.obj = list(), field, query = NULL, cutoff_frequency = NULL,
  low_freq_operator = NULL, minimum_should_match = NULL)
}
\arguments{
\item{index}{Index name}

\item{type}{Type name, Default: NULL, so all types}

\item{what}{In \code{\link{index}}, whether to get mappings, aliases, or settings.}

\item{...}{Further args passed on}

\item{x}{Input to various functions}

\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}
}
\description{
The elastic DSL
}
\details{
The DSL for \code{elastic} makes it easy to do queries against an Elasticsearch
instance, either local or remote.

The workflow with the DSL:
\enumerate{
 \item Start with the index to use, e.g., \code{index("shakespeare")}
 \item Define queries, e.g., \code{range(speech_number == 5, line_id > 3)}
 \item Execute search, e.g., \code{Search()}
}

Alternatively, if nothing follows a query definition, \code{\link{Search}} is called
to execute the search with the query as given. In a sense, this is essentially like
what \code{\link{dplyr}} does.
}
\examples{
\dontrun{
# Define index
index("shakespeare")
index("shakespeare", "scene")
index("shakespeare", "act")
index("plos")
index("gbif")

# DSL queries default to search across all indices
bool(must_not = list(term=list(speaker="KING HENRY IV")))

# bool query
bool(must_not = list(term=list(speaker="KING HENRY IV")))
index("shakespeare") \%>\%
   bool(must_not = list(term=list(speaker="KING HENRY IV")))
index("shakespeare") \%>\%
   bool(should = list(list(term=list(speech_number=1))))

# range query
index("shakespeare") \%>\% range( speech_number <= 5 )
index("shakespeare") \%>\% range( speech_number <= c(1,5) ) # doens't work
index("shakespeare") \%>\% range( speech_number >= c(1,5) ) # doens't work

# geographic query
## point
index("geoshape") \%>\%
   geoshape(field = "location", type = "envelope", coordinates = list(c(-30, 50), c(30, 0)))
# circle and radius
index("geoshape") \%>\%
   geoshape(field = "location", type = "circle", radius = "2000km",
            coordinates = c(-10, 45)) \%>\%
   n()
index("geoshape") \%>\%
   geoshape(field = "location", type = "circle", radius = "5000km",
            coordinates = c(-10, 45)) \%>\%
   n()
# polygon
coords <- list(c(80.0, -20.0), c(-80.0, -20.0), c(-80.0, 60.0), c(40.0, 60.0), c(80.0, -20.0))
index("geoshape") \%>\%
   geoshape(field = "location", type = "polygon", coordinates = coords)

# limit query
filter()
}
}

