% Generated by roxygen2 (4.0.2): do not edit by hand
\name{es_analyze}
\alias{es_analyze}
\title{Performs the analysis process on a text and return the tokens breakdown of the text.}
\usage{
es_analyze(text = NULL, field = NULL, index = NULL, analyzer = NULL,
  tokenizer = NULL, filters = NULL, char_filters = NULL,
  callopts = list())
}
\arguments{
\item{text}{The text on which the analysis should be performed (when request body is not used)}

\item{field}{Use the analyzer configured for this field (instead of passing the analyzer name)}

\item{index}{The name of the index to scope the operation}

\item{analyzer}{The name of the analyzer to use}

\item{tokenizer}{The name of the tokenizer to use for the analysis}

\item{filters}{A comma-separated list of filters to use for the analysis}

\item{char_filters}{A comma-separated list of character filters to use for the analysis}

\item{callopts}{Curl args passed on to httr::POST.}
}
\description{
Performs the analysis process on a text and return the tokens breakdown of the text.
}
\details{
This method can accept a sting of text in the body, but this function passes it as a
parameter in a GET request to simplify.
}
\examples{
\donttest{
es_analyze(text = 'this is a test', analyzer='standard')
es_analyze(text = 'this is a test', analyzer='whitespace')
es_analyze(text = 'this is a test', analyzer='stop')
es_analyze(text = 'this is a test', tokenizer='keyword', filters='lowercase')
es_analyze(text = 'this is a test', tokenizer='keyword', filters='lowercase',
   char_filters='html_strip')
es_analyze(text = 'this is a test', index = 'plos')
es_analyze(text = 'this is a test', index = 'shakespeare')
es_analyze(text = 'this is a test', index = 'shakespeare', callopts=verbose())
}
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/indices-analyze.html}
}

