% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{es_auth}
\alias{es_auth}
\title{Set authentication details}
\usage{
es_auth(es_base = NULL, es_port = NULL, es_user = NULL, es_pwd = NULL,
  force = FALSE)
}
\arguments{
\item{es_base}{(character) Base url}

\item{es_port}{(character) Port}

\item{es_user}{(character) User name}

\item{es_pwd}{(character) Password}

\item{force}{(logical) Force update}
}
\description{
Set authentication details
}
\keyword{internal}

