% Generated by roxygen2 (4.0.1): do not edit by hand
\name{es_cdbriver_auth}
\alias{es_cdbriver_auth}
\title{Authentication for CouchDB database using Elasticsearch river plugin.}
\usage{
es_cdbriver_auth(user = NULL, pwd = NULL, dbname,
  endpoint = "http://localhost", port = 9200)
}
\arguments{
\item{user}{Username}

\item{pwd}{Password}

\item{dbname}{Database name. (charcter)}

\item{endpoint}{the endpoint, defaults to localhost (http://127.0.0.1)}

\item{port}{port to connect to, defaults to 9200}
}
\description{
Authentication for CouchDB database using Elasticsearch river plugin.
}
\examples{
\dontrun{
es_cdbriver_auth(user="foo", pwd="bar")
}
}

