% Generated by roxygen2 (4.0.2): do not edit by hand
\name{es_clear_cache}
\alias{es_clear_cache}
\title{Clear either all caches or specific cached associated with one ore more indices.}
\usage{
es_clear_cache(index = NULL, filter = FALSE, filter_keys = NULL,
  fielddata = FALSE, query_cache = FALSE, id_cache = FALSE,
  callopts = list())
}
\arguments{
\item{index}{The name of the index to scope the operation}

\item{filter}{(logical) Clear filter caches}

\item{filter_keys}{(character) A vector of keys to clear when using the \code{filter_cache}
parameter (default: all)}

\item{fielddata}{(logical) Clear field data}

\item{query_cache}{(logical) Clear query caches}

\item{id_cache}{(logical) Clear ID caches for parent/child}

\item{callopts}{Curl args passed on to httr::POST.}
}
\description{
Clear either all caches or specific cached associated with one ore more indices.
}
\examples{
\donttest{
es_clear_cache()
es_clear_cache(index = "plos")
es_clear_cache(index = "shakespeare")
es_clear_cache(index = c("plos","shakespeare"))
es_clear_cache(filter = TRUE)
library('httr')
es_clear_cache(callopts=verbose())
}
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/indices-analyze.html}
}

