% Generated by roxygen2 (4.0.2): do not edit by hand
\name{es_cluster}
\alias{es_cluster}
\title{Elasticsearch cluster endpoints}
\usage{
es_cluster(what = "health", index = NULL, raw = FALSE,
  callopts = list(), verbose = TRUE, ...)
}
\arguments{
\item{what}{Which cluster endpiont to use, one of health, state, stats, reroute, settings, or
pending_tasks.}

\item{index}{Index}

\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then raw JSON.}

\item{callopts}{Curl args passed on to httr::POST.}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{...}{Further args passed on to elastic search HTTP API as parameters.}

\item{type}{Document type}
}
\description{
Elasticsearch cluster endpoints
}
\examples{
\dontrun{
es_cluster(what='health')
es_cluster('state')
es_cluster('settings')
es_cluster('stats')
es_cluster('pending_tasks')

# raw json data
es_cluster('health', raw=TRUE)
}
}

