% Generated by roxygen2 (4.0.2): do not edit by hand
\name{es_count}
\alias{es_count}
\title{Get counts of the number of records per index.}
\usage{
es_count(index = NULL, type = NULL, callopts = list(), verbose = TRUE,
  ...)
}
\arguments{
\item{index}{Index, defaults to all indices}

\item{type}{Document type}

\item{callopts}{Curl args passed on to httr::GET.}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{...}{Further args passed on to elastic search HTTP API as parameters.}
}
\description{
Get counts of the number of records per index.
}
\details{
There are a lot of terms you can use for Elasticsearch. See here
   \url{http://www.elasticsearch.org/guide/reference/query-dsl/} for the documentation.
}
\examples{
\dontrun{
es_count()
es_count(index='mran')
es_count(index='mran', type='metadata')
es_count(index='twitter')
}
}

