% Generated by roxygen2 (4.0.2): do not edit by hand
\name{es_index_delete}
\alias{es_index_delete}
\title{Start or stop indexing a document or many documents.}
\usage{
es_index_delete(index, raw = FALSE, callopts = list(), verbose = TRUE)
}
\arguments{
\item{index}{Index name}

\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then raw JSON.}

\item{callopts}{Curl args passed on to httr::POST.}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{...}{Further args passed on to elastic search HTTP API as parameters.}
}
\description{
Start or stop indexing a document or many documents.
}
\details{
The function returns TRUE. Though note that this can result even
   if the database does not exist in CouchDB.
}
\examples{
\dontrun{
es_index_delete(index='plos')
}
}
\references{
See docs for the Elasticsearch River plugin \url{#} that lets you
    easily index CouchDB databases.
}

