% Generated by roxygen2 (4.0.2): do not edit by hand
\name{es_ping}
\alias{es_ping}
\title{Ping an Elasticsearch server.}
\usage{
es_ping(raw = FALSE, callopts = list())
}
\arguments{
\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then raw JSON.}

\item{callopts}{Curl args passed on to httr::POST.}
}
\description{
Ping an Elasticsearch server.
}
\examples{
\dontrun{
es_ping()
}
}

