% Generated by roxygen2 (4.0.1): do not edit by hand
\name{es_status}
\alias{es_status}
\title{Get a list of your indices, just the names.}
\usage{
es_status(callopts = list(), raw = FALSE, verbose = TRUE, ...)
}
\description{
See \code{es_status} for more details on indices.
}
\examples{
\dontrun{
es_status()
}
}

