% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsl.R
\name{get_map}
\alias{get_map}
\title{Get mapping for verifying fields requested in DSL}
\usage{
get_map(index, type = NULL, ...)
}
\description{
Get mapping for verifying fields requested in DSL
}
\examples{
\dontrun{
shake <- get_map("shakespeare")
shake$line
# field names
names(shake$line$properties)
# field types
pluck(shake$line$properties, "type", "")
shake$scene
shake$act
}
}

