% Generated by roxygen2 (4.0.2): do not edit by hand
\name{elastic_GET}
\alias{elastic_GET}
\alias{elastic_POST}
\title{GET wrapper}
\usage{
elastic_GET(path, index = NULL, type = NULL, metric = NULL, node = NULL,
  clazz = NULL, raw, callopts, ...)

elastic_POST(path, index = NULL, type = NULL, clazz = NULL, raw, callopts,
  query, ...)
}
\arguments{
\item{path}{Elasticsearch API endpoint path}

\item{index}{Elasticsearch index}

\item{type}{Elasticsearch type}

\item{metric}{A metric to get}

\item{node}{The node}

\item{clazz}{Class to outupt}

\item{raw}{Raw JSON results as string}

\item{callopts}{Curl options}

\item{...}{Further args passed to Elasticsearch}

\item{query}{Query list or string to be passed on to body of the \code{POST} call.}

\item{path}{Elasticsearch API endpoint path}

\item{index}{Elasticsearch index}

\item{type}{Elasticsearch type}

\item{clazz}{Class to outupt}

\item{raw}{Raw JSON results as string}

\item{callopts}{Curl options}

\item{...}{Further args passed to Elasticsearch. If passed in as list, converted to JSON, if
passed in as character, validate the json (using \code{jsonlite::validate}), then pass in
to request.}
}
\description{
GET wrapper

POST wrapper
}
\keyword{internal}

