% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Get the basic info from the current cluster}
\usage{
info(...)
}
\arguments{
\item{...}{Further args passed on to print for the es_conn class.}
}
\description{
Get the basic info from the current cluster
}
\examples{
\dontrun{
connect()
info()
}
}
