% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nodes}
\alias{nodes}
\alias{nodes_hot_threads}
\alias{nodes_info}
\alias{nodes_stats}
\title{Elasticsearch nodes endpoints.}
\usage{
nodes_stats(node = NULL, metric = NULL, raw = FALSE, fields = NULL,
  verbose = TRUE, ...)

nodes_info(node = NULL, metric = NULL, raw = FALSE, verbose = TRUE, ...)

nodes_hot_threads(node = NULL, metric = NULL, threads = 3,
  interval = "500ms", type = NULL, raw = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{node}{The node}

\item{metric}{A metric to get}

\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then raw JSON.}

\item{fields}{You can get information about field data memory usage on node level or on
index level}

\item{verbose}{If TRUE (default) the url call used printed to console}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}

\item{threads}{(character) Number of hot threads to provide. Default: 3}

\item{interval}{(character) The interval to do the second sampling of threads. Default: 500ms}

\item{type}{(character) The type to sample, defaults to cpu, but supports wait and block to
see hot threads that are in wait or block state.}
}
\description{
Elasticsearch nodes endpoints.
}
\details{
\url{http://bit.ly/11gezop}

By default, all stats are returned. You can limit this by combining any of indices, os, process,
jvm, network, transport, http, fs, breaker and thread_pool. With the metric parameter you can
select zero or more of:

\itemize{
 \item indices Indices stats about size, document count, indexing and deletion times, search
 times, field cache size, merges and flushes
 \item fs File system information, data path, free disk space, read/write stats
 \item http HTTP connection information
 \item jvm JVM stats, memory pool information, garbage collection, buffer pools
 \item network TCP information
 \item os Operating system stats, load average, cpu, mem, swap
 \item process Process statistics, memory consumption, cpu usage, open file descriptors
 \item thread_pool Statistics about each thread pool, including current size, queue and rejected
 tasks
 \item transport Transport statistics about sent and received bytes in cluster communication
 \item breaker Statistics about the field data circuit breaker
}
}
\examples{
\donttest{
(out <- nodes_stats())
nodes_stats(node = names(out$nodes))
nodes_stats(metric='get')
nodes_stats(metric='jvm')
nodes_stats(metric=c('os','process'))
nodes_info()
nodes_info(metric='process')

# these may not work
nodes_shutdown()
nodes_hot_threads()
}
}

