% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ping.r
\name{ping}
\alias{ping}
\title{Ping an Elasticsearch server.}
\usage{
ping(...)
}
\arguments{
\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Ping an Elasticsearch server.
}
\examples{
\donttest{
ping()
}
}

