% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsl.R
\name{range}
\alias{range}
\title{Define range query for the DSL}
\usage{
range(.obj = list(), ..., boost = 1, cache = FALSE)
}
\description{
Define range query for the DSL
}
\examples{
\dontrun{
range( speech_number == 5, line_id > 3 )
range( speech_number <= c(1,5) )
range( speech_number >= c(1,5) )
}
}

