% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search}
\alias{search}
\title{Full text search of Elasticsearch - URI requests}
\usage{
search(index = NULL, type = NULL, raw = FALSE, callopts = list(), ...)
}
\arguments{
\item{index}{Index}

\item{type}{Document type}

\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then raw JSON.}

\item{callopts}{Curl args passed on to httr::POST.}

\item{...}{Further args passed on to elastic search HTTP API as parameters.}
}
\description{
Full text search of Elasticsearch - URI requests
}
\examples{
\dontrun{
search(index="twitter")
search(index="twitter", type="tweet")
search(index="twitter", type="mention")
search(index="twitter", type="tweet", q="what")
search(index="twitter", type="tweet", sort="message")

res <- search(index="mran", q="snapshotDate>='2013-10-01'")
res$hits$total

search(index="shakespeare", q="what")
res <- search(index="shakespeare", q="snapshotDate>='2013-10-01'")
search(index="shakespeare", q="createdTime>='2013-10-01'")
search(index="shakespeare", size=3, explain=TRUE)

# Get raw data
search(index="twitter", type="tweet", raw=TRUE)

# Curl debugging
library('httr')
search(index="twitter", type="tweet", callopts=verbose())

search(index="shakespeare", size=0, aggs = )
}
}

