% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termvectors.R
\name{termvectors}
\alias{termvectors}
\title{Termvectors}
\usage{
termvectors(index, type, id = NULL, body = list(), pretty = TRUE,
  field_statistics = TRUE, fields = NULL, offsets = TRUE, parent = NULL,
  payloads = TRUE, positions = TRUE, realtime = TRUE,
  preference = "random", routing = NULL, term_statistics = FALSE,
  version = NULL, version_type = NULL, ...)
}
\arguments{
\item{index}{(character) The index in which the document resides.}

\item{type}{(character) The type of the document.}

\item{id}{(character) The id of the document, when not specified a doc
param should be supplied.}

\item{body}{(character) Define parameters and or supply a document to get
termvectors for}

\item{pretty}{(logical) pretty print. Default: \code{TRUE}}

\item{field_statistics}{(character) Specifies if document count, sum
of document frequencies and sum of total term frequencies should be
returned. Default: \code{TRUE}}

\item{fields}{(character) A comma-separated list of fields to return.}

\item{offsets}{(character) Specifies if term offsets should be returned.
Default: \code{TRUE}}

\item{parent}{(character) Parent id of documents.}

\item{payloads}{(character) Specifies if term payloads should be returned.
Default: \code{TRUE}}

\item{positions}{(character) Specifies if term positions should be returned.
Default: \code{TRUE}}

\item{realtime}{(character) Specifies if request is real-time as opposed to
near-real-time (Default: \code{TRUE}).}

\item{preference}{(character) Specify the node or shard the operation
should be performed on (Default: \code{random}).}

\item{routing}{(character) Specific routing value.}

\item{term_statistics}{(character) Specifies if total term frequency and
document frequency should be returned. Default: \code{FALSE}}

\item{version}{(character) Explicit version number for concurrency control}

\item{version_type}{(character) Specific version type, valid choices are:
'internal', 'external', 'external_gte', 'force'}

\item{...}{Curl args passed on to \code{\link[httr]{POST}}}
}
\description{
Termvectors
}
\details{
Returns information and statistics on terms in the fields of a
particular document. The document could be stored in the index or
artificially provided by the user (Added in 1.4). Note that for
documents stored in the index, this is a near realtime API as the term
vectors are not available until the next refresh.
}
\examples{
\dontrun{
connect()
if (!index_exists('plos')) {
  plosdat <- system.file("examples", "plos_data.json", package = "elastic")
  invisible(docs_bulk(plosdat))
}
if (!index_exists('omdb')) {
  omdb <- system.file("examples", "omdb.json", package = "elastic")
  invisible(docs_bulk(omdb))
}

body <- '{
  "fields" : ["title"],
  "offsets" : true,
  "positions" : true,
  "term_statistics" : true,
  "field_statistics" : true
}'
termvectors('plos', 'article', 29, body = body)

body <- '{
  "fields" : ["Plot"],
  "offsets" : true,
  "positions" : true,
  "term_statistics" : true,
  "field_statistics" : true
}'
termvectors('omdb', 'omdb', 'AVXdx8Eqg_0Z_tpMDyP_', body = body)
}
}
\references{
\url{http://www.elastic.co/guide/en/elasticsearch/reference/current/docs-termvectors.html}
}
