% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpdocs.R
\name{units-time}
\alias{units-time}
\title{Time units}
\description{
Whenever durations need to be specified, eg for a timeout parameter, the duration can 
be specified as a whole number representing time in milliseconds, or as a time value 
like 2d for 2 days. The supported units are:
}
\details{
\tabular{ll}{ 
y \tab Year \cr
M \tab Month \cr
w \tab Week \cr
d \tab Day \cr
h \tab Hour \cr
m \tab Minute \cr
s \tab Second \cr
}
}
\seealso{
\code{\link{units-distance}}
}
