% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\alias{connection}
\title{Set connection details to an Elasticsearch engine.}
\usage{
connect(es_host = "127.0.0.1", es_port = 9200,
  es_transport_schema = "http", es_user = NULL, es_pwd = NULL,
  force = FALSE, errors = "simple", es_base = NULL, ...)

connection()
}
\arguments{
\item{es_host}{(character) The base host, defaults to \code{127.0.0.1}}

\item{es_port}{(character) port to connect to, defaults to \code{9200} (optional)}

\item{es_transport_schema}{(character) http or https. Default: \code{http}}

\item{es_user}{(character) User name, if required for the connection. You can specify, 
but ignored for now.}

\item{es_pwd}{(character) Password, if required for the connection. You can specify, but
ignored for now.}

\item{force}{(logical) Force re-load of connection details}

\item{errors}{(character) One of simple (Default) or complete. Simple gives http code and 
error message on an error, while complete gives both http code and error message, 
and stack trace, if available.}

\item{es_base}{(character) Deprecated, use \code{es_host}}

\item{...}{Further args passed on to print for the es_conn class.}
}
\description{
Set connection details to an Elasticsearch engine.
}
\details{
The default configuration is set up for localhost access on port 9200,
with no username or password.

\code{\link{connection}} does not ping the Elasticsearch server, but only
prints connection details.

On package load, \code{\link{connect}} is run to set the default base url and port.

Internally, we store your preferences with environment variables. That means you 
can set your env vars permanently in .Renviron file, and use them on a server e.g., 
as private env vars.
}
\examples{
\dontrun{
# the default is set to localhost and port 9200
connect()

# or set to a different base url
# connect('162.243.152.53')

# See connection details
connection()
}
}

