% Generated by roxygen2 (4.0.2): do not edit by hand
\name{connect}
\alias{connect}
\title{Set connection details to an Elasticsearch engine.}
\usage{
connect(es_base = "http://127.0.0.1", es_port = 9200, es_user = NULL,
  es_pwd = NULL, es_key = NULL, force = FALSE, ...)
}
\arguments{
\item{es_base}{The base url, defaults to localhost (http://127.0.0.1)}

\item{es_port}{port to connect to, defaults to 5984}

\item{es_user}{User name, if required for the connection. You can specify, but
ignored for now.}

\item{es_pwd}{Password, if required for the connection. You can specify, but
ignored for now.}

\item{es_key}{An API key}

\item{force}{Force re-authorization.}

\item{...}{Further args passed on to print for the es_conn class.}
}
\description{
Set connection details to an Elasticsearch engine.
}
\details{
The default configuration is set up for localhost access on port 9200,
with no username or password.

Pass on the returned 'es_conn' object to other functions in this package.
}
\examples{
\dontrun{
# the default is set to localhost and port 9200
connect()

# or set to a different base url
connect('http://162.243.152.56')
}
}

