% Generated by roxygen2 (4.0.2): do not edit by hand
\name{docs_bulk}
\alias{docs_bulk}
\title{Use the bulk API to create, index, update, or delete documents.}
\usage{
docs_bulk(filename, raw = FALSE, callopts = list())
}
\arguments{
\item{filename}{Path to a file to load in the bulk API}

\item{raw}{(logical) Get raw JSON back or not.}

\item{callopts}{Pass on options to POST call.}
}
\description{
Use the bulk API to create, index, update, or delete documents.
}
\details{
More on the Bulk API:
   \url{http://www.elasticsearch.org/guide/en/elasticsearch/guide/current/bulk.html}.
}
\examples{
\donttest{
plosdat <- system.file("examples", "plos_data.json", package = "elastic")
docs_bulk(file=plosdat)
aliases_get()
index_delete(index='plos')
aliases_get()
}
}

