% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docs_delete.R
\name{docs_delete}
\alias{docs_delete}
\title{Delete a document.}
\usage{
docs_delete(index, type, id, refresh = NULL, routing = NULL,
  timeout = NULL, version = NULL, version_type = NULL,
  callopts = list(), ...)
}
\arguments{
\item{index}{(character) The name of the index. Required}

\item{type}{(character) The type of the document. Required}

\item{id}{(numeric) The document ID. Required}

\item{refresh}{(logical) Refresh the index after performing the operation}

\item{routing}{(charcter) Specific routing value}

\item{timeout}{(charcter) Explicit operation timeout, e.g,. 5m (for 5 minutes)}

\item{version}{(character) Explicit version number for concurrency control}

\item{version_type}{(character) Specific version type. One of internal or external}

\item{callopts}{Curl args passed on to \code{\link[httr]{DELETE}}}

\item{...}{Further args to query DSL}
}
\description{
Delete a document.
}
\examples{
\donttest{
docs_delete(index='plos', type='article', id=35)
docs_get(index='plos', type='article', id=35) # and the document is gone
}
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/docs-delete.html}
}

