% Generated by roxygen2 (4.0.2): do not edit by hand
\name{docs_mget}
\alias{docs_mget}
\title{Get multiple documents via the multiple get API.}
\usage{
docs_mget(index = NULL, type = NULL, id = NULL, type_id = NULL,
  index_type_id = NULL, source = NULL, fields = NULL, raw = FALSE,
  callopts = list(), verbose = TRUE, ...)
}
\arguments{
\item{index}{Index}

\item{type}{Document type}

\item{id}{Document id}

\item{type_id}{List of vectors of length 2, each with an element for type and id.}

\item{index_type_id}{List of vectors of length 3, each with an element for index,
type, and id.}

\item{source}{XXX}

\item{fields}{Fields to return from the response object.}

\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then raw JSON.}

\item{callopts}{Curl args passed on to httr::POST.}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{...}{Further args passed on to elastic search HTTP API as parameters.}
}
\description{
Get multiple documents via the multiple get API.
}
\details{
There are a lot of terms you can use for Elasticsearch. See here
   \url{http://www.elasticsearch.org/guide/reference/query-dsl/} for the documentation.
}
\examples{
\donttest{
# Same index and type
docs_mget(index="shakespeare", type="line", id=c(9,10))
tmp <- docs_mget(index="mran", type="metadata", id=c('plyr','ggplot2'), raw=TRUE)
es_parse(tmp)
docs_mget(index="mran", type="metadata", id=c('plyr','ggplot2'), fields='description')
docs_mget(index="mran", type="metadata", id=c('plyr','ggplot2'), source=TRUE)

library("httr")
docs_mget(index="twitter", type="tweet", id=1:2, callopts=verbose())

# Same index, but different types
docs_mget(index="shakespeare", type_id=list(c("scene",1), c("line",20)))
docs_mget(index="twitter", type_id=list(c("tweet",1), c("mention",2)))
docs_mget(index="twitter", type_id=list(c("tweet",1), c("mention",2)), fields='user')
docs_mget(index="twitter", type_id=list(c("tweet",1), c("mention",2)), fields=c('user','message'))

# Different indeces and different types
# pass in separately
docs_mget(index_type_id=list(c("shakespeare","line",1), c("appdotnet","share",1)))
docs_mget(index_type_id=list(c("twitter","mention",1), c("appdotnet","share",1)))
}
}

