% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{elastic-defunct}
\alias{elastic-defunct}
\title{Defunct functions in elastic}
\description{
\itemize{
 \item \code{\link{mlt}}: The MLT API has been removed, use More Like This Query 
 via \code{\link{Search}}
 \item \code{\link{nodes_shutdown}}: The _shutdown API has been removed. Instead, 
 setup Elasticsearch to run as a service (see Running as a Service on Linux 
 (\url{https://www.elastic.co/guide/en/elasticsearch/reference/2.0/setup-service.html}) or 
 Running as a Service on Windows 
 (\url{https://www.elastic.co/guide/en/elasticsearch/reference/2.0/setup-service-win.html})) 
 or use the -p command line option to write the PID to a file.
 \item \code{\link{index_status}}: _status route for the index API has been removed. 
 Replaced with the Indices Stats and Indices Recovery APIs.
 \item \code{\link{mapping_delete}}: Elasticsearch dropped this route in their API. Instead 
 of deleting a mapping, delete the index and recreate with a new mapping.
}
}
