% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsl.R
\name{elasticdsl}
\alias{bool}
\alias{bool_}
\alias{elasticdsl}
\alias{index}
\alias{print.index}
\alias{range}
\alias{range_}
\title{The elastic DSL}
\usage{
index(index, type = NULL, what = "mappings", ...)

\method{print}{index}(x, ...)

range(.obj = list(), ..., boost = 1, time_zone = NULL, execution = NULL,
  cache = FALSE)

range_(.obj = list(), ..., .dots)

bool(.obj = list(), ...)

bool_(.obj = list(), ..., .dots)
}
\arguments{
\item{index}{asdfds}

\item{type}{asdfds}

\item{what}{asdfds}

\item{...}{adfadsf}

\item{x}{adfadf}

\item{.obj}{asdfds}
}
\description{
The elastic DSL
}
\details{
The DSL for \code{elastic} makes it easy to do queries against an Elasticsearch
instance, either local or remote.

The workflow with the DSL:
\enumerate{
 \item Start with the index to use, e.g., \code{index("shakespeare")}
 \item Define queries, e.g., \code{range(speech_number == 5, line_id > 3)}
 \item Execute search, e.g., \code{Search()}
}

Alternatively, if nothing follows a query definition, \code{\link{Search}} is called
to execute the search with the query as given. In a sense, this is essentially like
what \code{\link{dplyr}} does.
}
\examples{
\dontrun{
# Get mapping for verifying fields requested in DSL
index("shakespeare")
index("shakespeare", "scene")
index("shakespeare", "act")
index("plos")
index("gbif")

# Define bool query
bool(must_not = list(term=list(speaker="KING HENRY IV")))
index("shakespeare") \%>\%
   bool(must = list(term=list(speaker="KING HENRY IV")))
index("shakespeare") \%>\%
   bool(must_not = list(term=list(speaker="KING HENRY IV")))
index("shakespeare") \%>\%
   bool(should = list(list(term=list(speech_number=1))))

bool(must = list(term=list(user="kimchy")),
     must_not = list(term=list(user="kimchy")))

# Define range query
index("shakespeare") \%>\% range( speech_number <= 5 )
index("shakespeare") \%>\% range( speech_number <= c(1,5) )
range( speech_number >= c(1,5) )
}
}

