% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsl.R
\name{elasticdsl}
\alias{bool}
\alias{elasticdsl}
\alias{index}
\alias{print.index}
\alias{range}
\title{The elastic DSL}
\usage{
index(index, type = NULL, what = "mappings", ...)

\method{print}{index}(x, ...)

range(.obj = list(), ..., boost = 1, cache = FALSE)

bool(.obj = list(), ...)
}
\arguments{
\item{index}{asdfds}

\item{type}{asdfds}

\item{what}{asdfds}

\item{...}{adfadsf}

\item{x}{adfadf}

\item{.obj}{asdfds}
}
\description{
The elastic DSL
}
\details{
The DSL for \code{elastic} makes it easy to do queries against an Elasticsearch
instance, either local or remote.

The workflow with the DSL:
\enumerate{
 \item Start with the index to use, e.g., \code{index("shakespeare")}
 \item Define queries, e.g., \code{range(speech_number == 5, line_id > 3)}
 \item Execute search, e.g., \code{Search()}
}

Alternatively, if nothing follows a query definition, \code{\link{Search}} is called
to execute the search with the query as given. In a sense, this is essentially like
what \code{\link{dplyr}} does.
}
\examples{
\dontrun{
# Get mapping for verifying fields requested in DSL
index("shakespeare")
index("plos")
index("gbif")

# Define bool query for the DSL
index("shakespeare") \%>\%
   bool(must = list(term=list(user="kimchy")))
bool(must_not = list(term=list(user="kimchy")))
bool(should = list(list(term=list(tag="sometag")), list(term=list(tag="someothertag"))))

bool(must = list(term=list(user="kimchy")),
     must_not = list(term=list(user="kimchy")))

# Define range query for the DSL
range( speech_number == 5, line_id > 3 )
range( speech_number <= c(1,5) )
range( speech_number >= c(1,5) )
}
}

