% Generated by roxygen2 (4.0.2): do not edit by hand
\name{es_auth}
\alias{es_auth}
\title{Set authentication details}
\usage{
es_auth(es_base = NULL, es_port = NULL, es_user = NULL, es_pwd = NULL,
  es_key = NULL, force = FALSE)
}
\arguments{
\item{es_base}{(character) Base url}

\item{es_port}{(character) Port}

\item{es_user}{(character) User name}

\item{es_pwd}{(character) Password}

\item{es_key}{(character) API key}

\item{force}{(logical) Force update}
}
\description{
Only base url and port and used right now. Will add use or username, password, key, etc. later.
}
\details{
\itemize{
 \item You can enter your details using the client_id and api_key parameters directly.
 \item You can execute the function without any inputs. The function then first looks in your
 options for the option variables digocean_client_id and digocean_api_key. If they are not found
 the function asks you to enter them. You can set force=TRUE to force the function to ask
 you for new id and key.
 \item Set your options using the function \code{options}. See examples.
 \item Set your options in your .Rprofile file with entries
 \code{options(es_base = '<clientid>')}, \code{options(es_port = '<port>')},
 \code{options(es_user = '<port>')}, \code{options(es_pwd = '<port>')}, and
 \code{options(es_key = '<port>')}
}
}
\keyword{internal}

