% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{es_auth}
\alias{es_auth}
\title{Set authentication details}
\usage{
es_auth(es_host = NULL, es_port = NULL, es_path = NULL,
  es_transport_schema = NULL, es_user = NULL, es_pwd = NULL,
  force = FALSE, es_base = NULL)
}
\arguments{
\item{es_host}{(character) Base url}

\item{es_port}{(character) Port}

\item{es_transport_schema}{(character) http or https. Default: \code{http}}

\item{es_user}{(character) User name}

\item{es_pwd}{(character) Password}

\item{force}{(logical) Force update}

\item{es_base}{(character) deprecated, use es_host}
}
\description{
Set authentication details
}
\keyword{internal}
