% Generated by roxygen2 (4.0.1): do not edit by hand
\name{es_bulk}
\alias{es_bulk}
\title{Use the bulk API to insert or delete documents.}
\usage{
es_bulk(conn, index = NULL, type = NULL, id = NULL, type_id = NULL,
  index_type_id = NULL, source = NULL, fields = NULL, raw = FALSE,
  callopts = list(), verbose = TRUE, ...)
}
\arguments{
\item{index}{Index}

\item{type}{Document type}

\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then raw JSON.}

\item{callopts}{Curl args passed on to httr::POST.}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{...}{Further args passed on to elastic search HTTP API as parameters.}

\item{id}{Document id}

\item{source}{XXX}

\item{fields}{Fields to return from the response object.}

\item{type_id}{List of vectors of length 2, each with an element for type and id.}

\item{index_type_id}{List of vectors of length 3, each with an element for index,
type, and id.}
}
\description{
Use the bulk API to insert or delete documents.
}
\details{
There are a lot of terms you can use for Elasticsearch. See here
   \url{http://www.elasticsearch.org/guide/reference/query-dsl/} for the documentation.
}
\examples{
\dontrun{
init <- es_connect()
es_bulk(init, index="twitter", type="tweet", id=1:2)
}
}

