% Generated by roxygen2 (4.0.2): do not edit by hand
\name{es_bulk}
\alias{es_bulk}
\alias{make_bulk_plos}
\title{Use the bulk API to create, index, update, or delete documents.}
\usage{
es_bulk(filename, raw = FALSE, callopts = list())

make_bulk_plos(n = 1000, filename = "~/plos_data.json")
}
\arguments{
\item{filename}{Path to a file to load in the bulk API}

\item{raw}{(logical) Get raw JSON back or not.}

\item{callopts}{Pass on options to POST call.}

\item{n}{Number of documents to get from PLOS API.}
}
\description{
Use the bulk API to create, index, update, or delete documents.
}
\details{
More on the Bulk API:
   \url{http://www.elasticsearch.org/guide/en/elasticsearch/guide/current/bulk.html}.
}
\examples{
\donttest{
make_bulk_plos(1000, filename = "~/plos_data.json")
es_bulk(file="~/plos_data.json")
es_aliases()
es_index_delete(index='plos')
es_aliases()
}
}

