% Generated by roxygen2 (4.0.1): do not edit by hand
\name{es_count}
\alias{es_count}
\title{Execute a query and get the number of matches for that query.}
\usage{
es_count(index = NULL, type = NULL, callopts = list(), verbose = TRUE,
  ...)
}
\arguments{
\item{index}{Index}

\item{type}{Document type}

\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then raw JSON.}

\item{callopts}{Curl args passed on to httr::POST.}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{...}{Further args passed on to elastic search HTTP API as parameters.}

\item{id}{Document id}

\item{source}{XXX}

\item{fields}{Fields to return from the response object.}

\item{exists}{XXX}
}
\description{
Execute a query and get the number of matches for that query.
}
\details{
There are a lot of terms you can use for Elasticsearch. See here
   \url{http://www.elasticsearch.org/guide/reference/query-dsl/} for the documentation.
}
\examples{
\dontrun{
es_count(index='mran')
es_count(index='twitter')
}
}

