% Generated by roxygen2 (4.0.1): do not edit by hand
\name{es_GET}
\alias{es_GET}
\title{Wrapper around httr::GET for this package}
\usage{
es_GET(path, index = NULL, type = NULL, clazz = NULL, raw, callopts, ...)
}
\arguments{
\item{path}{Elasticsearch API endpoint path}

\item{index}{Elasticsearch index}

\item{type}{Elasticsearch type}

\item{clazz}{Class to outupt}

\item{raw}{Raw JSON results as string}

\item{callopts}{Curl options}

\item{...}{Further args passed to Elasticsearch}
}
\description{
Wrapper around httr::GET for this package
}

