% Generated by roxygen2 (4.0.2): do not edit by hand
\name{es_index}
\alias{es_index}
\title{Start or stop indexing a document or many documents.}
\usage{
es_index(index = NULL, type = NULL, id = NULL, source = FALSE,
  fields = NULL, exists = FALSE, raw = FALSE, callopts = list(),
  verbose = TRUE, ...)
}
\arguments{
\item{index}{Index name}

\item{type}{Document type}

\item{id}{Document id}

\item{source}{Ignored for now}

\item{fields}{Fields to add.}

\item{exists}{(logical) Exists or not}

\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then raw JSON.}

\item{callopts}{Curl args passed on to httr::POST.}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{...}{Further args passed on to elastic search HTTP API as parameters.}
}
\description{
Start or stop indexing a document or many documents.
}
\details{
The function returns TRUE. Though note that this can result even
   if the database does not exist in CouchDB.
}
\examples{
\dontrun{
es_index(index='twitter', type='tweet', id=10)

es_index(index='things', type='tweet', id=10)
}
}
\references{
See docs for the Elasticsearch River plugin \url{#} that lets you
    easily index CouchDB databases.
}

