% Generated by roxygen2 (4.0.2): do not edit by hand
\name{es_search}
\alias{es_search}
\title{Full text search of Elasticsearch - URI requests}
\usage{
es_search(index = NULL, type = NULL, raw = FALSE, callopts = list(),
  ...)
}
\arguments{
\item{index}{Index}

\item{type}{Document type}

\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then raw JSON.}

\item{callopts}{Curl args passed on to httr::POST.}

\item{...}{Further args passed on to elastic search HTTP API as parameters.}
}
\description{
Full text search of Elasticsearch - URI requests
}
\examples{
\dontrun{
es_search(index="twitter")
es_search(index="twitter", type="tweet")
es_search(index="twitter", type="mention")
es_search(index="twitter", type="tweet", q="what")
es_search(index="twitter", type="tweet", sort="message")

res <- es_search(index="mran", q="snapshotDate>='2013-10-01'")
res$hits$total

es_search(index="shakespeare", q="what")
res <- es_search(index="shakespeare", q="snapshotDate>='2013-10-01'")
es_search(index="shakespeare", q="createdTime>='2013-10-01'")
es_search(index="shakespeare", size=3, explain=TRUE)

# Get raw data
es_search(index="twitter", type="tweet", raw=TRUE)

# Curl debugging
library('httr')
es_search(index="twitter", type="tweet", callopts=verbose())

es_search(index="shakespeare", size=0, aggs = )
}
}

