% Generated by roxygen2 (4.0.2): do not edit by hand
\name{es_status}
\alias{es_status}
\title{Get status details for your cluster.}
\usage{
es_status(callopts = list(), raw = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{callopts}{Curl args passed on to httr::GET.}

\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then raw JSON.}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{...}{Further args passed on to elastic search HTTP API as parameters.}
}
\description{
Get status details for your cluster.
}
\examples{
\dontrun{
es_status()
}
}

