% Generated by roxygen2 (4.0.2): do not edit by hand
\name{index}
\alias{es_index_close}
\alias{es_index_create}
\alias{es_index_delete}
\alias{es_index_exists}
\alias{es_index_get}
\alias{es_index_open}
\alias{index}
\title{Elasticsearch indices APIs}
\usage{
es_index_get(index = NULL, features = NULL, raw = FALSE,
  callopts = list(), verbose = TRUE, ...)

es_index_exists(index, callopts = list())

es_index_delete(index, raw = FALSE, callopts = list(), verbose = TRUE)

es_index_create(index = NULL, type = NULL, id = NULL, fields = NULL,
  raw = FALSE, callopts = list(), verbose = TRUE, ...)

es_index_close(index, callopts = list())

es_index_open(index, callopts = list())
}
\arguments{
\item{index}{A comma-separated list of index names}

\item{features}{A comma-separated list of features. One or more of settings, mappings,
warmers or aliases}

\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then raw JSON.}

\item{callopts}{Curl args passed on to httr::POST.}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{...}{Further args passed on to elastic search HTTP API as parameters.}

\item{type}{Document type}

\item{id}{Document id}

\item{fields}{Fields to add.}
}
\description{
Elasticsearch indices APIs
}
\examples{
\donttest{
# get information on an index
es_index_get(index='shakespeare')
es_index_get(index='shakespeare', features=c('settings','mappings'))
es_index_get(index='shakespeare', features='aliases')
es_index_get(index='shakespeare', features='warmers')

# check for index existence
es_index_exists(index='shakespeare')
es_index_exists(index='plos')

# delete an index
es_index_delete(index='plos')

# create an index
es_index_create(index='twitter', type='tweet', id=10)
es_index_create(index='things', type='tweet', id=10)

# close an index
es_index_close('plos')

# open an index
es_index_open('plos')
}
}
\author{
Scott Chamberlain <myrmecocystus@gmail.com>
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/indices.html}
}

