% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest}
\alias{ingest}
\alias{pipeline_create}
\alias{pipeline_get}
\alias{pipeline_delete}
\alias{pipeline_simulate}
\title{Ingest API operations}
\usage{
pipeline_create(id, body, ...)

pipeline_get(id, filter_path = NULL, ...)

pipeline_delete(id, body, ...)

pipeline_simulate(body, id = NULL, ...)
}
\arguments{
\item{id}{(character) one or more pipeline id's. with delete, you can use
a wildcard match}

\item{body}{body describing pipeline, see examples and Elasticsearch docs}

\item{...}{Curl args passed on to \code{\link[httr:POST]{httr::POST()}}, \code{\link[httr:GET]{httr::GET()}},
\code{\link[httr:PUT]{httr::PUT()}}, or \code{\link[httr:DELETE]{httr::DELETE()}}}

\item{filter_path}{(character) fields to return. deafults to all if not given}
}
\value{
a named list
}
\description{
Ingest API operations
}
\details{
ingest/pipeline functions available in Elasticsearch v5 and greater
}
\examples{
\dontrun{
# create
body1 <- '{
  "description" : "do a thing",
  "version" : 123,
  "processors" : [
    {
      "set" : {
        "field": "foo",
        "value": "bar"
      }
    }
  ]
}'
body2 <- '{
  "description" : "do another thing",
  "processors" : [
    {
      "set" : {
        "field": "stuff",
        "value": "things"
      }
    }
  ]
}'
pipeline_create(id = 'foo', body = body1)
pipeline_create(id = 'bar', body = body2)

# get
pipeline_get(id = 'foo')
pipeline_get(id = 'bar')
pipeline_get(id = 'foo', filter_path = "*.version")
pipeline_get(id = c('foo', 'bar')) # get >1

# delete
pipeline_delete(id = 'foo')

# simulate
## with pipeline included
body <- '{
  "pipeline" : {
    "description" : "do another thing",
    "processors" : [
      {
        "set" : {
          "field": "stuff",
          "value": "things"
        }
      }
    ]
  },
  "docs" : [
    { "_source": {"foo": "bar"} },
    { "_source": {"foo": "world"} }
  ]
}'
pipeline_simulate(body)

## referencing existing pipeline
body <- '{
  "docs" : [
    { "_source": {"foo": "bar"} },
    { "_source": {"foo": "world"} }
  ]
}'
pipeline_simulate(body, id = "foo")
}
}
\references{
\url{https://www.elastic.co/guide/en/elasticsearch/reference/current/ingest-apis.html}
}
