% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mapping}
\alias{es_field_mapping_get}
\alias{es_mapping_create}
\alias{es_mapping_delete}
\alias{es_mapping_get}
\alias{es_type_exists}
\alias{mapping}
\title{Mapping management}
\usage{
es_mapping_create(index, type, body, ...)

es_mapping_delete(index, type, ...)

es_mapping_get(index = NULL, type = NULL, ...)

es_field_mapping_get(index = NULL, type = NULL, field,
  include_defaults = FALSE, ...)

es_type_exists(index, type, ...)
}
\arguments{
\item{index}{(character) An index}

\item{type}{(character) A document type}

\item{body}{(list) Either a list or json, representing the query.}

\item{...}{Options passed on to \code{\link[httr]{HEAD}}}

\item{field}{(character) One or more field names}

\item{include_defaults}{(logical) Whether to return default values}
}
\description{
Mapping management
}
\details{
\itemize{
 \item es_mapping_create - http://bit.ly/1xbWqFo
 \item es_type_exists - http://bit.ly/10HkZvH
 \item es_mapping_delete - http://bit.ly/10Mmvgi
 \item es_mapping_get - http://bit.ly/1AN2oiw
 \item es_field_mapping_get - http://bit.ly/1wHKgCA
}

\stong{NOTE} For the delete method, Elasticsearch documentation notes that: "Note, most times,
it make more sense to reindex the data into a fresh index compared to delete large chunks of it."
}
\examples{
\donttes{
# Used to check if a type/types exists in an index/indices
es_type_exists(index = "plos", type = "article")
es_type_exists(index = "plos", type = "articles")
es_type_exists(index = "shakespeare", type = "line")

# The put mapping API allows to register specific mapping definition for a specific type.
## a good mapping body
body <- list(citation = list(properties = list(
 journal = list(type="string"),
 year = list(type="long")
)))
es_mapping_create(index = "plos", type = "citation", body=body)

### or as json
body <- '{
  "citation": {
    "properties": {
      "journal": { "type": "string" },
      "year": { "type": "long" }
}}}'
es_mapping_delete("plos", "citation")
es_mapping_create(index = "plos", type = "citation", body=body)
es_mapping_get("plos", "citation")

## A bad mapping body
body <- list(things = list(properties = list(
  journal = list("string")
)))
es_mapping_create(index = "plos", type = "things", body=body)

# Delete a mapping
es_mapping_delete("plos", "citation")

# Get mappings
es_mapping_get('_all')
es_mapping_get(index = "plos")
es_mapping_get(index = c("shakespeare","plos"))
es_mapping_get(index = "shakespeare", type = "act")
es_mapping_get(index = "shakespeare", type = c("act","line"))

# Get field mappings
es_field_mapping_get(index = "_all", type=c('article','line'), field = "text")
es_field_mapping_get(index = "plos", type = "article", field = "title")
es_field_mapping_get(index = "plos", type = "article", field = "*")
es_field_mapping_get(index = "plos", type = "article", field = "title", include_defaults = TRUE)
es_field_mapping_get(type = c("article","record"), field = c("title","class"))
es_field_mapping_get(type = "a*", field = "t*")
}
}

