% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.r
\name{ping}
\alias{ping}
\title{Ping an Elasticsearch server.}
\usage{
ping(...)
}
\arguments{
\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Ping an Elasticsearch server.
}
\examples{
\dontrun{
ping()
}
}
\seealso{
\code{\link{connect}}
}
