% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchapis}
\alias{searchapis}
\title{Search functions.}
\description{
Search functions.
}
\details{
Elasticsearch search APIs include the following functions:
\itemize{
 \item \code{\link{Search}} - Search using the URI search API or pass in query
 specification in the body of the request.
 \item \code{\link{search_shards}} - Search shards.
 \item \code{\link{mlt}} - More like this queries.
 \item \code{\link{count}} - Get counts for various searches.
 \item \code{\link{explain}} - Computes a score explanation for a query and a specific
 document. This can give useful feedback whether a document matches or didn't match
 a specific query.
}

More will be added soon.
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search.html}
}

