% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchapis}
\alias{searchapis}
\title{Search functions.}
\description{
Search functions.
}
\details{
Elasticsearch search APIs include the following functions:
\itemize{
 \item \code{\link{es_search}} - Search using the URI search API
 \item \code{\link{search_body}} - Pass in search specification in request body
 \item \code{\link{mlt}} - More like this queries
 \item \code{\link{count}} - Get counts for various searches
}

More will be added soon.
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search.html}
}

