% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{tasks}
\alias{tasks}
\alias{tasks_cancel}
\title{Elasticsearch tasks endpoints}
\usage{
tasks(task_id = NULL, nodes = NULL, actions = NULL,
  parent_task_id = NULL, detailed = FALSE, group_by = NULL,
  wait_for_completion = FALSE, timeout = NULL, raw = FALSE, ...)

tasks_cancel(node_id = NULL, task_id = NULL, nodes = NULL,
  actions = NULL, parent_task_id = NULL, detailed = FALSE,
  group_by = NULL, wait_for_completion = FALSE, timeout = NULL,
  raw = FALSE, ...)
}
\arguments{
\item{task_id}{a task id}

\item{nodes}{(character) The nodes}

\item{actions}{(character) Actions}

\item{parent_task_id}{(character) A parent task ID}

\item{detailed}{(character) get detailed results. Default: FALSE}

\item{group_by}{(character) "nodes" (default, i.e., NULL) or "parents"}

\item{wait_for_completion}{(logical) wait for completion. Default: FALSE}

\item{timeout}{(integer) timeout time}

\item{raw}{If TRUE (default), data is parsed to list. If FALSE, then
raw JSON.}

\item{...}{Curl args passed on to \code{\link[httr]{GET}} or 
\code{\link[httr]{POST}}}

\item{node_id}{a node id}
}
\description{
Elasticsearch tasks endpoints
}
\examples{
\dontrun{
connect()

tasks()
# tasks(parent_task_id = "1234")

# delete a task
# tasks_cancel()
}
}
\references{
\url{https://www.elastic.co/guide/en/elasticsearch/reference/current/tasks.html}
}
